Organize license feature: action, checker, page; Update docs; Add tests
