+++ includes/index.php
@@ -1,6 +1,5 @@
 <?php
 
-require_once __DIR__ . '/license.php';
-require_once __DIR__ . '/license-checker.php';
+require_once __DIR__ . '/license/index.php';
 require_once __DIR__ . '/register.php';
 require_once __DIR__ . '/legacy.php';
